+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "mapviewer",
-  "version": "1.2.2",
+  "version": "3.0.0",
   "private": true,
   "scripts": {
     "serve:kloosters": "vue-cli-service serve src/kloosters.js",
@@ -22,6 +22,7 @@
     "mobile-device-detect": "^0.4.3",
     "ol": "^6.4.3",
     "ol-ext": "^3.1.14",
+    "openlayers-ext": "^3.1.5",
     "proj4": "^2.6.2",
     "uuid": "^8.3.0",
     "vue": "^2.6.11",
