+++ src/helpers/ViewerConfig.js
@@ -7,6 +7,7 @@ import ViewerServiceGeoJSON from "./service/ViewerServiceGeoJSON";
 import ViewerServiceKloostersByYear from "./kloosters/service/ViewerServiceKloostersByYear";
 import ViewerServiceKloostersAll from "./kloosters/service/ViewerServiceKloostersAll";
 import ViewerServiceKloostersSingle from "./kloosters/service/ViewerServiceKloostersSingle";
+import ViewerServiceKerken from "./kerken/service/ViewerServiceKerken";
 
 class ViewerConfig {
   constructor() {
@@ -63,6 +64,7 @@ class ViewerConfig {
     if (service_config.type === 'kloosters_by_year') return new ViewerServiceKloostersByYear(service_config);
     if (service_config.type === 'kloosters_all') return new ViewerServiceKloostersAll(service_config);
     if (service_config.type === 'kloosters_single') return new ViewerServiceKloostersSingle(service_config);
+    if (service_config.type === 'kerken') return new ViewerServiceKerken(service_config);
   }
 }
